requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.12.3
asttokens==2.0.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
awscli==1.18.69
backports.entry-points-selectable==1.1.0
base58==2.1.0
bcrypt==3.1.7
bitarray==1.2.2
black==21.9b0
blinker==1.4
boto3==1.20.2
botocore==1.23.2
Brlapi==0.7.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cytoolz==0.11.0
dataclassy==0.11.1
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
eip712==0.1.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.1
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
fasteners==0.14.1
filelock==3.3.1
future==0.18.2
hexbytes==0.2.2
httplib2==0.14.0
hypothesis==6.24.0
idna==3.3
importlib-metadata==1.5.0
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jedi==0.18.1
jmespath==0.9.4
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
lxml==4.6.4
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onelogin==2.0.2
onelogin-aws-assume-role==1.8.1
packaging==21.0
paramiko==2.6.0
parsimonious==0.8.1
parso==0.8.2
pathspec==0.9.0
pexpect==4.6.0
Pillow==7.0.0
pipx==0.16.4
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.21
protobuf==3.19.0
psutil==5.8.0
ptpython==3.0.20
py==1.10.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.8.0
pycryptodome==3.11.0
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.3
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
pyte==0.8.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.16.0
pythx==1.6.1
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.4.1
regex==2021.10.23
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
rlp==2.0.1
roman==2.0.0
rsa==4.0
s3transfer==0.5.0
SecretStorage==2.3.1
semantic-version==2.8.5
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sortedcontainers==2.4.0
systemd-python==234
toml==0.10.2
tomli==1.2.2
toolz==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
userpath==1.7.0
varint==1.0.2
virtualenv==20.10.0
vvm==0.1.0
vyper==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
web3==5.24.0
websockets==9.1
wfuzz==2.4.5
wrapt==1.13.2
xkit==0.0.0
yarl==1.7.0
youtube-dl==2020.3.24
zipp==1.0.0
